+++ SensingKit/SKBatteryStatus.m
@@ -32,14 +32,16 @@
 {
     if (self = [super init])
     {
-        // Register for battery level and state change notifications.
+        // Register for battery level and state change notifications
         [[NSNotificationCenter defaultCenter] addObserver:self
-                                                 selector:@selector(batteryLevelChanged:)
+                                                 selector:@selector(batterySensorStateChanged:)
                                                      name:UIDeviceBatteryLevelDidChangeNotification object:nil];
         
         [[NSNotificationCenter defaultCenter] addObserver:self
-                                                 selector:@selector(batteryStateChanged:)
+                                                 selector:@selector(batterySensorStateChanged:)
                                                      name:UIDeviceBatteryStateDidChangeNotification object:nil];
+        
+        // powerModeChanged: will be registered later (on startSensing:)
         
         self.configuration = configuration;
     }
@@ -80,8 +82,8 @@
         if (error) {
             
             NSDictionary *userInfo = @{
-                                       NSLocalizedDescriptionKey: NSLocalizedString(@"Battery sensor is not available.", nil),
-                                       };
+                NSLocalizedDescriptionKey: NSLocalizedString(@"Battery Status sensor is not available.", nil),
+            };
             
             *error = [NSError errorWithDomain:SKErrorDomain
                                          code:SKSensorNotAvailableError
@@ -90,14 +92,27 @@
         return NO;
     }
     
+    // start receiving battery level and state change notifications
     [UIDevice currentDevice].batteryMonitoringEnabled = YES;
+    
+    // start receiving powerModeChanged related notifications
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(batterySensorStateChanged:)
+                                                 name:NSProcessInfoPowerStateDidChangeNotification
+                                               object:nil];
     
     return YES;
 }
 
 - (BOOL)stopSensing:(NSError **)error
 {
+    // stop receiving battery level and state change notifications
     [UIDevice currentDevice].batteryMonitoringEnabled = NO;
+    
+    // stop receiving powerModeChanged related notifications
+    [[NSNotificationCenter defaultCenter] removeObserver:self
+                                                    name:NSProcessInfoPowerStateDidChangeNotification
+                                                  object:nil];
     
     return [super stopSensing:error];
 }
@@ -112,18 +127,23 @@
     return [UIDevice currentDevice].batteryState;
 }
 
-- (void)batteryLevelChanged:(NSNotification *)notification
+- (SKLowPowerModeState)lowPowerModeState
 {
-    SKBatteryStatusData *data = [[SKBatteryStatusData alloc] initWithLevel:[self batteryLevel]
-                                                     withState:[self batteryState]];
-    
-    [self submitSensorData:data error:NULL];
+    if ([[NSProcessInfo processInfo] isLowPowerModeEnabled])
+    {
+        return SKLowPowerModeStateEnabled;
+    }
+    else
+    {
+        return SKLowPowerModeStateDisabled;
+    }
 }
 
-- (void)batteryStateChanged:(NSNotification *)notification
+- (void)batterySensorStateChanged:(NSNotification *)notification
 {
     SKBatteryStatusData *data = [[SKBatteryStatusData alloc] initWithLevel:[self batteryLevel]
-                                                     withState:[self batteryState]];
+                                                                 withState:[self batteryState]
+                                                         lowPowerModeState:[self lowPowerModeState]];
     
     [self submitSensorData:data error:NULL];
 }