+++ SensingKit/SKBatteryStatusData.h
@@ -30,6 +30,23 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
+ *  These constants indicate the type of the device's low power mode.
+ */
+typedef NS_ENUM(NSUInteger, SKLowPowerModeState) {
+    
+    /**
+     *  Disabled. This is the normal mode that all device's features are enabled.
+     */
+    SKLowPowerModeStateDisabled = 0,
+    
+    /**
+     *  When a device is in Low Power Mode, some features are disabled or take longer to be completed. However, the device lasts longer.
+     */
+    SKLowPowerModeStateEnabled,
+};
+
+
+/**
  *  An instance of SKBatteryStatusData encapsulates measurements related to the Battery Status sensor.
  */
 @interface SKBatteryStatusData : SKSensorData
@@ -47,7 +64,9 @@ NS_ASSUME_NONNULL_BEGIN
  *
  *  @return An SKBatteryStatusData object.
  */
-- (instancetype)initWithLevel:(CGFloat)level withState:(UIDeviceBatteryState)state NS_DESIGNATED_INITIALIZER;
+- (instancetype)initWithLevel:(CGFloat)level
+                    withState:(UIDeviceBatteryState)state
+            lowPowerModeState:(SKLowPowerModeState)lowPowerModeState NS_DESIGNATED_INITIALIZER;
 
 /**
  *  A float number that indicates the current battery charge level. Value ranges from 0.0 (fully discharged) to 1.0 (fully charged).
@@ -63,6 +82,16 @@ NS_ASSUME_NONNULL_BEGIN
  *  A string value that describes the state of the battery, classified as Charging, Full, Unplugged or Unknown.
  */
 @property (nonatomic, readonly, copy) NSString *stateString;
+
+/**
+ *  An enumerator that descrived the state of the device's Low Power Mode, classified as Disabled or Enabled.
+ */
+@property (nonatomic, readonly) SKLowPowerModeState lowPowerModeState;
+
+/**
+ *  A string value that describes the state of the device's Low Power Mode, classified as Disabled or Enabled.
+ */
+@property (nonatomic, readonly, copy) NSString *lowPowerModeStateString;
 
 /**
  *  A string with a CSV formatted header that describes the data of the Battery Status sensor. This method is useful in combination with the csvString instance method of an SKSensorData object.